+++ lib/src/commands/do/upgrade/deps.dart
@@ -47,6 +47,7 @@ class UpgradeDepsCommand extends DirCommand<void> {
     required Directory directory,
     required GgLog ggLog,
     bool? majorVersions,
+    Map<String, dynamic> options = const {},
   }) => get(directory: directory, ggLog: ggLog, majorVersions: majorVersions);
 
   @override